Add GUC for TAM arrow cache size
The arrow cache is used to cache decompressed arrow segments and is
mostly useful for index scans that access values across segments in
non-segment order.

The fixed cache size was set too small for index scans on bigger data
sets. Therefore, increase the default size and make it configurable via
a GUC.

+++ tsl/src/hypercore/arrow_cache.h
@@ -53,7 +53,7 @@ typedef struct ArrowColumnCache
 	size_t arrow_column_cache_lru_count; /* Arrow column cache LRU list count */
 	dlist_head arrow_column_cache_lru;	 /* Arrow column cache LRU list */
 	HTAB *htab;							 /* Arrow column cache */
-	uint16 maxsize;
+	size_t maxsize;
 } ArrowColumnCache;
 
 typedef struct ArrowTupleTableSlot ArrowTupleTableSlot;
